fix: telegram multiformat (#89)
This PR adds:

- support for messages that contain special objects such as links,
phone, email, formatting. `message.content` can be in the form of a list
which raises the error when checking for type string
- support for chats with deleted accounts by adding the value "Deleted
Account" if `from` field is empty. This is an issue when trying to call
`map_ai_messages()` as it won't be able to determine the sender
resulting in no `AIMessage`

Reproducing the bug:

1. export a Telegram chat that contains links or other special
formatting like bold text in JSON format
2. import and run the function `TelegramChatLoader()` - it will throw
the error `Chat Loaders only support messages with content type string,
got ...`